feat(vault): migrate credential management to separate module
The changes in this commit move the credential management functionality from the `props` package to a new `auth` package. This is done to better separate concerns and make the codebase more modular.

The key changes are:

- Moved the `Credentials` struct and related methods from `pkg/vault/props/creds.go` to `pkg/vault/auth/creds.go`.
- Updated the `vault` package to use the new `auth.NewCredentials()` method instead of `props.NewCredentials()`.
- Removed the `pkg/auth/credential.go` file as its contents have been integrated into the new `auth` package.

This change will improve the maintainability and testability of the credential management functionality in the Vault.
