+++ pkg/vault/auth/authenticator.go
